Fix ok button on spacebar
Use the button callback for the ok button instead of button-1 events, so
the handler is called when the button is in focus and is pressed with
the spacebar.  Give the event parameter on the ok button handler a
default value so we can continue to accept the Enter events.